Account for shrinking holes during printing
When printing a hole with the classic extrusion 3D printers
the inner diameter usually comes out a bit smaller than designed.

Account for this by making the hole diameter 0.5 mm larger than the bolt
outer diameter.

This is probably bad idea as the value of 0.5 is rather arbitrary
and as now this design implicitly expect to be printed with
an extrusion printer.